Corrected the \to command
Return an error message when the user is not found on the Array. As explained on the notes:
"I'll associate user ID/name to the socket when implement nicknames so now, when a user disconnects he still on the array"
By now when connected it remains on the array after disconnected. I'm leaving it for later.